Lock deleted concurrent queue instead
