--- src/main/java/com/cardcostapi/infrastructure/ILock.java
@@ -1,5 +1,0 @@
-package com.cardcostapi.infrastructure;
-
-public interface ILock {
-    public Object aquire(String key);
-}
