--- src/main/java/com/cardcostapi/infrastructure/InMemoryLock.java
@@ -1,16 +1,0 @@
-package com.cardcostapi.infrastructure;
-
-import org.springframework.stereotype.Component;
-
-import java.util.concurrent.ConcurrentHashMap;
-
-@Component
-public class InMemoryLock implements ILock {
-
-    private final ConcurrentHashMap<String, Object> locks = new ConcurrentHashMap<>();
-
-    @Override
-    public Object aquire(String bin) {
-        return locks.computeIfAbsent(bin, k -> new Object());
-    }
-}
