+++ src/main/java/com/cardcostapi/services/IBinLookupService.java
@@ -1,5 +1,7 @@
 package com.cardcostapi.services;
 
+import java.util.concurrent.ExecutionException;
+
 public interface IBinLookupService {
     public String getCountryByBin(String pan);
 }
