+++ src/main/java/com/cardcostapi/services/InMemoryRateLimit.java
@@ -39,9 +39,7 @@ public class InMemoryRateLimit implements IRateLimitService {
         while (!timestamps.isEmpty() && timestamps.peek().isBefore(cutoff)) {
             timestamps.poll();
         }
-        if (timestamps.size() < limit == false) {
-            System.out.println("wtf!");
-        }
+
         return timestamps.size() < limit;
     }
 
@@ -54,6 +52,7 @@ public class InMemoryRateLimit implements IRateLimitService {
             accessMap.put(key, timestamps);
         }
         timestamps.offer(Instant.now());
+        System.out.println("Aumentando RL: " + key);
     }
 
     // Solo para testing
